modularized the webpack config creator functions
